bug: Timebase and sampling interval calculation
I'm still having problems with the sampling interval. The oscilloscope is now at the point where
 it does the full sweep. However, it only returns the first part of the data. Seems like the sma
pling rate is too fast so that the buffer is filled (according to max_samples) while it's still
sweeping.

Current change add a further constraint to GetTimebase by both having an enum sampling rate and pointer to a sampling rate.