bug: adjusting no samples collected
If the calculated sampling rate is faster than 10 MS/s (rated max), then we set is as the baseline. However, the oscilloscope doesn't automatically adjust the number of samples collected. That meant that it tended to collect samples for way too long at the highest frequency.

I added an adjustment to the sampling interval calculation to 'manually' account for this.